re-create notifications as to not bother the users
