+++ src/app/App.tsx
@@ -7,13 +7,9 @@ import {
 } from 'react-native';
 import {
   adaptNavigationTheme,
-  Button,
-  Dialog,
   MD3DarkTheme,
   MD3LightTheme,
   PaperProvider,
-  Portal,
-  Text,
 } from 'react-native-paper';
 import {
   DarkTheme as NavigationDarkTheme,
@@ -75,7 +71,6 @@ const VALID_PRODUCT_IDS = [
 
 // Migration key for AsyncStorage
 const REALM_MIGRATION_COMPLETE_KEY = 'ralph_realm_to_watermelon_migration_complete';
-const NOTIFICATIONS_REACTIVATION_KEY = 'ralph_notifications_reactivation_needed';
 
 
 // Initialize Stack Navigator
@@ -111,9 +106,7 @@ const App: React.FC<{
   // State
   const [handledInitialNotificationId, setHandledInitialNotificationId] =
     useState<string | undefined>();
-  const [showNotificationReactivationDialog, setShowNotificationReactivationDialog] =
-    useState<boolean>(false);
-
+  
   // Function to check and run migration if needed
   const checkAndRunMigration = async () => {
     try {
@@ -131,11 +124,6 @@ const App: React.FC<{
       if (result.success) {
         // Mark migration as complete in AsyncStorage
         await AsyncStorage.setItem(REALM_MIGRATION_COMPLETE_KEY, 'true');
-
-        // If user had notifications enabled, store flag to show dialog
-        if (result.hadNotifications) {
-          setShowNotificationReactivationDialog(true);
-        }
 
         console.log('Migration completed and marked as done.');
       } else {
@@ -330,22 +318,6 @@ const App: React.FC<{
       behavior={Platform.OS === 'ios' ? 'padding' : 'height'}>
       <DatabaseProvider>
         <PaperProvider theme={theme}>
-          {/* Notification Reactivation Dialog */}
-          <Portal>
-            <Dialog
-              visible={showNotificationReactivationDialog}
-              onDismiss={() => setShowNotificationReactivationDialog(false)}>
-              <Dialog.Title>{t('notifications:reactivation_needed')}</Dialog.Title>
-              <Dialog.Content>
-                <Text variant="bodyMedium">
-                  {t('notifications:reactivation_message')}
-                </Text>
-              </Dialog.Content>
-              <Dialog.Actions>
-                <Button onPress={() => setShowNotificationReactivationDialog(false)}>{t('buttons:close')}</Button>
-              </Dialog.Actions>
-            </Dialog>
-          </Portal>
 
           <NavigationContainer onStateChange={onNavigationStateChange}>
             <Stack.Navigator
